Refactor lib/validator
