+++ lib/metadata/virtual_container_host.go
@@ -17,7 +17,6 @@ package metadata
 import (
 	"crypto/tls"
 	"errors"
-	"fmt"
 	"net/mail"
 	"net/url"
 	"time"
